requirements.txt
asgiref==3.8.1
asttokens==3.0.0
decorator==5.2.1
Django==4.2.21
django-debug-toolbar==5.1.0
django-extensions==4.1
exceptiongroup==1.2.2
executing==2.2.0
ipython==8.18.1
jedi==0.19.2
matplotlib-inline==0.1.7
parso==0.8.4
pexpect==4.9.0
pillow==11.2.1
prompt_toolkit==3.0.50
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.1
sqlparse==0.5.3
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.13.2
wcwidth==0.2.13
celery==5.3.6
redis==5.0.1
django-celery-beat==2.8.0
django-environ==0.11.2
djangorestframework==3.16.0
django-filter==25.1
djangorestframework-simplejwt==5.5.0
psycopg2-binary==2.9.9
